+++ R/EpiNow-package.R
@@ -1,0 +1,18 @@
+#' The 'EpiNow' package.
+#'
+#' @description To identify changes in the reproduction number, rate of spread, 
+#' and doubling time during the course of outbreaks whilst accounting 
+#' potential biases due to delays in case reporting.
+#'
+#' @docType package
+#' @name EpiNow-package
+#' @aliases EpiNow
+#' @useDynLib EpiNow, .registration = TRUE
+#' @import methods
+#' @import Rcpp
+#' @importFrom rstan sampling
+#'
+#' @references
+#' Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
+#'
+NULL