+++ NEWS.md
@@ -1,3 +1,14 @@
+
+# EpiNow 0.2.0 
+
+* Time-varying window for Rt estimation.
+* Refactored all code in `data.table`.
+* Refactored code base.
+* Added mapping from onset cases to infection with upscaling to account for additional truncation.
+* Added minimal example to all functions
+* Added an approximate mapping for reporting delay that has an adaptive case threshold.
+* Switched to using a generation time estimate rather than a serial interval (see `data-raw/ganyani-generation-time.rs)
+
 # EpiNow 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.